Refactored some of Scrutinizer into token
Added pyliter.token.StyledToken which is a duck type for
tokenize.TokenInfo. StyledToken supports style dictionaries and saves
the raw string and allows the string attribute to be updated (unlike
TokenInfo which is a namedtuple). Expanded the exact types range of
types to include things like KEYWORD, BUILTIN, DOCSTRING, FNAME,
CNAME, INAME, LVAL, SELF and more.

Lastly, added a ANSIStyledToken subclass of StyledToken that overrides
the string property to apply the style dictionary to the string. The
idea here is to support future StyledToken subclasses that do not
rely on ANSI escape codes to style the text.