Render text with OpenGL
All the text coloring with ANSI escape codes was thrown out and
replaced with rendering the text into an OpenGL window. The idea
is to specify a python text file and option line number to start
on and number of lines.  The original idea was to use head and
tail on text files and render those fragments, but that messed
up tokenizing the file.  I haven't figured out how to render
to an off-screen framebuffer yet, so for now the text is
rendered to a visible window and saves a PNG to hard-coded
file 'output.png' in the current working directory. Pressing
escape will dismiss the window.

Summary of changes:

- refactored pyliter.token extensively
- pyliter command-line got a slew of new options
- added module pyliter.render to handle on-screen text rendering
- added module pyliter.document to provide a PythonDocument
- added classmethod pyliter.color.Color.from_any
- updated pyliter/resources/default_style.yaml with new OpenGL attributes